Add Ansible Development Tools and Environment features (#17)
* Add Ansible Development Tools feature

This commit adds the Ansible Development Tools feature to the project. It includes a new README.md file with information about the feature and its usage. It also includes a devcontainer-feature.json file that defines the feature and its options. Additionally, it includes an install.sh script and a library_scripts.sh file that are used to install and configure the feature. Finally, it includes a scenarios.json file and a test.sh script for testing the feature.

* Add Ansible Development Environment feature

This commit adds the Ansible Development Environment feature to the project. It includes a README.md file with instructions and options for using the feature, a devcontainer-feature.json file that defines the feature, an install.sh script to install the feature, a library_scripts.sh file with utility functions, and test files for the feature.

* Add ansible-dev-tools and ansible-development-environment to the test workflow

* Update version number and fix typo in devcontainer feature